Refactor: Separate main.dart into multiple files and widgets
