+++ lib/widgets/bottom_bar_button.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class BottomBarButton extends StatelessWidget {
+  const BottomBarButton({super.key, required this.icon, this.onPressed});
+
+  final IconData icon;
+  final VoidCallback? onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return IconButton(icon: Icon(icon), onPressed: onPressed);
+  }
+}
